streamlit_app/requirements.txt
streamlit
requests
python-dotenv
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn
statsmodels
